[ALPS09346881] disp: Fix coverity - compare unsigned to 0
[Description]
Fix coverity -
   compare unsigned to 0

[Test]
MTBF

MTK-Commit-Id: 156e9b0ce4bf3f4505240d7b4fc6c5800f99836e

CR-Id: ALPS09346881
Feature: display_mobile

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_drm_drv.c
@@ -2171,7 +2171,7 @@ static int mtk_atomic_commit(struct drm_device *drm,
 			return 0;
 		}
 
-		CRTC_MMP_MARK(drm_crtc_index(crtc), mml_job_status, mtk_crtc->is_mml,
+		CRTC_MMP_MARK((int)drm_crtc_index(crtc), mml_job_status, mtk_crtc->is_mml,
 			atomic_read(&mtk_crtc->wait_mml_last_job_is_flushed));
 
 		if (mtk_crtc->is_mml) {
